Don't setup credentials if they are empty
This is particularly important to allow dependabot to work with private
dependencies.  A PR created by dependabot won't have access to the
secrets, so the inputs will be there but empty, causing to create an
credentials file with no username or password, which ends up in an auth
error.

To use private repos as dependencies, they must be configured in the
organization.

For more details see:
* https://github.com/dependabot/dependabot-core/issues/3253#issuecomment-797125425
* https://docs.github.com/en/organizations/keeping-your-organization-secure/managing-security-settings-for-your-organization/managing-security-and-analysis-settings-for-your-organization#allowing-dependabot-to-access-private-dependencies

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>